+++ src/app/(home)/videos/[videoId]/page.tsx
@@ -11,6 +11,7 @@ const VideoIdPage = async ({ params }: VideoIdPageProps) => {
   const { videoId } = await params;
 
   void trpc.videos.getOne.prefetch({ id: videoId });
+  void trpc.comments.getMany.prefetch({ videoId });
 
   return (
     <HydrateClient>